requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
google-generativeai==0.3.2 